refactor: error handling routes

+++ src/repository/client-repository.js
@@ -1,11 +1,12 @@
 import connection from '../sql/connection.js';
 import Client from '../model/client-model.js';
+import { BadRequestError, NotFoundError } from '../../utils/api-errors.js';
 
 class ClientRepository {
   async getAllClients() {
     const sql = `SELECT * FROM agenda`;
     return new Promise((resolve, reject) => {
-      const { rows } = connection.query(sql, (error, result) => {
+      connection.query(sql, (error, result) => {
         if (error) {
           reject(error);
         } else {
@@ -18,7 +19,16 @@ class ClientRepository {
   async createClient(Client) {
     const existingClient = await this.getClientByName(Client.nameClient);
     if (existingClient && existingClient.length > 0)
-      throw new Error('Client already exists');
+      throw new BadRequestError('Client already exists');
+
+    if (
+      !Client.nameClient ||
+      !Client.contact ||
+      !Client.proceidure ||
+      !Client.deadline ||
+      !Client.price
+    )
+      throw new BadRequestError('Missing required fields');
 
     const sql = `INSERT INTO agenda (nameClient, contact, proceidure, deadline, price) VALUES (?, ?, ?, ?, ?)`;
     const values = [
@@ -28,6 +38,7 @@ class ClientRepository {
       Client.deadline,
       Client.price,
     ];
+
     return new Promise((resolve, reject) => {
       connection.query(sql, values, (error, result) => {
         if (error) {
@@ -42,7 +53,8 @@ class ClientRepository {
   async updateClient(id, Client) {
     const existingClient = await this.getClientById(id);
     if (!existingClient || existingClient.length === 0)
-      throw new Error('Client not found');
+      throw new NotFoundError('Client not found');
+
     const sql = `UPDATE agenda SET nameClient = ?, contact = ?, proceidure = ?, deadline = ?, price = ? WHERE id = ?`;
     const values = [
       Client.nameClient,
@@ -64,6 +76,10 @@ class ClientRepository {
   }
 
   async deleteClient(id) {
+    const existingClient = await this.getClientById(id);
+    if (!existingClient || existingClient.length === 0)
+      throw new NotFoundError('Client not found');
+
     const sql = `DELETE FROM agenda WHERE id = ?`;
     const values = [id];
     return new Promise((resolve, reject) => {
